+++ flash/pointcloud/segmentation/data.py
@@ -2,7 +2,8 @@ from typing import Any, Callable, Dict, Optional, Tuple
 
 from flash.core.data.data_module import DataModule
 from flash.core.data.data_source import DataSource, DefaultDataKeys, DefaultDataSources
-from flash.core.data.process import Deserializer, Preprocess
+from flash.core.data.io.input_transform import InputTransform
+from flash.core.data.process import Deserializer
 from flash.core.utilities.imports import requires
 from flash.pointcloud.segmentation.open3d_ml.sequences_dataset import SequencesDataset
 
@@ -52,7 +53,7 @@ class PointCloudSegmentationFoldersDataSource(DataSource):
         }
 
 
-class PointCloudSegmentationPreprocess(Preprocess):
+class PointCloudSegmentationInputTransform(InputTransform):
     def __init__(
         self,
         train_transform: Optional[Dict[str, Callable]] = None,
@@ -90,4 +91,4 @@ class PointCloudSegmentationPreprocess(Preprocess):
 
 class PointCloudSegmentationData(DataModule):
 
-    preprocess_cls = PointCloudSegmentationPreprocess
+    input_transform_cls = PointCloudSegmentationInputTransform