+++ tests/core/data/test_callbacks.py
@@ -18,7 +18,7 @@ from torch import tensor
 
 from flash.core.data.callback import BaseDataFetcher
 from flash.core.data.data_module import DataModule
-from flash.core.data.process import DefaultPreprocess
+from flash.core.data.io.input_transform import DefaultInputTransform
 from flash.core.utilities.stages import RunningStage
 
 
@@ -43,7 +43,7 @@ def test_base_data_fetcher(tmpdir):
         @classmethod
         def from_inputs(cls, train_data: Any, val_data: Any, test_data: Any, predict_data: Any) -> "CustomDataModule":
 
-            preprocess = DefaultPreprocess()
+            input_transform = DefaultInputTransform()
 
             return cls.from_data_source(
                 "default",
@@ -51,7 +51,7 @@ def test_base_data_fetcher(tmpdir):
                 val_data=val_data,
                 test_data=test_data,
                 predict_data=predict_data,
-                preprocess=preprocess,
+                input_transform=input_transform,
                 batch_size=5,
             )