+++ tests/graph/classification/test_model.py
@@ -19,12 +19,12 @@ import torch
 from flash import Trainer
 from flash.__main__ import main
 from flash.core.data.data_pipeline import DataPipeline
-from flash.core.utilities.imports import _TORCH_GEOMETRIC_AVAILABLE
+from flash.core.utilities.imports import _GRAPH_AVAILABLE
 from flash.graph.classification import GraphClassifier
-from flash.graph.classification.data import GraphClassificationPreprocess
+from flash.graph.classification.data import GraphClassificationInputTransform
 from tests.helpers.utils import _GRAPH_TESTING
 
-if _TORCH_GEOMETRIC_AVAILABLE:
+if _GRAPH_AVAILABLE:
     from torch_geometric import datasets
 
 
@@ -40,7 +40,7 @@ def test_train(tmpdir):
     """Tests that the model can be trained on a pytorch geometric dataset."""
     tudataset = datasets.TUDataset(root=tmpdir, name="KKI")
     model = GraphClassifier(num_features=tudataset.num_features, num_classes=tudataset.num_classes)
-    model.data_pipeline = DataPipeline(preprocess=GraphClassificationPreprocess())
+    model.data_pipeline = DataPipeline(input_transform=GraphClassificationInputTransform())
     train_dl = torch.utils.data.DataLoader(tudataset, batch_size=4)
     trainer = Trainer(default_root_dir=tmpdir, fast_dev_run=True)
     trainer.fit(model, train_dl)
@@ -51,7 +51,7 @@ def test_val(tmpdir):
     """Tests that the model can be validated on a pytorch geometric dataset."""
     tudataset = datasets.TUDataset(root=tmpdir, name="KKI")
     model = GraphClassifier(num_features=tudataset.num_features, num_classes=tudataset.num_classes)
-    model.data_pipeline = DataPipeline(preprocess=GraphClassificationPreprocess())
+    model.data_pipeline = DataPipeline(input_transform=GraphClassificationInputTransform())
     val_dl = torch.utils.data.DataLoader(tudataset, batch_size=4)
     trainer = Trainer(default_root_dir=tmpdir, fast_dev_run=True)
     trainer.validate(model, val_dl)
@@ -62,7 +62,7 @@ def test_test(tmpdir):
     """Tests that the model can be tested on a pytorch geometric dataset."""
     tudataset = datasets.TUDataset(root=tmpdir, name="KKI")
     model = GraphClassifier(num_features=tudataset.num_features, num_classes=tudataset.num_classes)
-    model.data_pipeline = DataPipeline(preprocess=GraphClassificationPreprocess())
+    model.data_pipeline = DataPipeline(input_transform=GraphClassificationInputTransform())
     test_dl = torch.utils.data.DataLoader(tudataset, batch_size=4)
     trainer = Trainer(default_root_dir=tmpdir, fast_dev_run=True)
     trainer.test(model, test_dl)
@@ -73,7 +73,7 @@ def test_predict_dataset(tmpdir):
     """Tests that we can generate predictions from a pytorch geometric dataset."""
     tudataset = datasets.TUDataset(root=tmpdir, name="KKI")
     model = GraphClassifier(num_features=tudataset.num_features, num_classes=tudataset.num_classes)
-    data_pipe = DataPipeline(preprocess=GraphClassificationPreprocess())
+    data_pipe = DataPipeline(input_transform=GraphClassificationInputTransform())
     out = model.predict(tudataset, data_source="datasets", data_pipeline=data_pipe)
     assert isinstance(out[0], int)