+++ tests/image/segmentation/test_model.py
@@ -26,7 +26,7 @@ from flash.core.data.data_pipeline import DataPipeline
 from flash.core.data.data_source import DefaultDataKeys
 from flash.core.utilities.imports import _IMAGE_AVAILABLE
 from flash.image import SemanticSegmentation
-from flash.image.segmentation.data import SemanticSegmentationPreprocess
+from flash.image.segmentation.data import SemanticSegmentationInputTransform
 from tests.helpers.utils import _IMAGE_TESTING, _SERVE_TESTING
 
 # ======== Mock functions ========
@@ -106,7 +106,7 @@ def test_unfreeze():
 def test_predict_tensor():
     img = torch.rand(1, 3, 64, 64)
     model = SemanticSegmentation(2, backbone="mobilenetv3_large_100")
-    data_pipe = DataPipeline(preprocess=SemanticSegmentationPreprocess(num_classes=1))
+    data_pipe = DataPipeline(input_transform=SemanticSegmentationInputTransform(num_classes=1))
     out = model.predict(img, data_source="tensors", data_pipeline=data_pipe)
     assert isinstance(out[0], list)
     assert len(out[0]) == 64
@@ -117,7 +117,7 @@ def test_predict_tensor():
 def test_predict_numpy():
     img = np.ones((1, 3, 64, 64))
     model = SemanticSegmentation(2, backbone="mobilenetv3_large_100")
-    data_pipe = DataPipeline(preprocess=SemanticSegmentationPreprocess(num_classes=1))
+    data_pipe = DataPipeline(input_transform=SemanticSegmentationInputTransform(num_classes=1))
     out = model.predict(img, data_source="numpy", data_pipeline=data_pipe)
     assert isinstance(out[0], list)
     assert len(out[0]) == 64
@@ -146,8 +146,8 @@ def test_jit(tmpdir, jitter, args):
 @mock.patch("flash._IS_TESTING", True)
 def test_serve():
     model = SemanticSegmentation(2)
-    # TODO: Currently only servable once a preprocess has been attached
-    model._preprocess = SemanticSegmentationPreprocess()
+    # TODO: Currently only servable once a input_transform has been attached
+    model._input_transform = SemanticSegmentationInputTransform()
     model.eval()
     model.serve()